fix 32-bit system compiler warning

+++ src/store/redis/rdsstore.c
@@ -1066,7 +1066,7 @@ static void redis_subscriber_callback(redisAsyncContext *c, void *r, void *privd
       unsigned    chid_present = 0;
       ngx_str_t   extracted_channel_id;
       unsigned    msgbuf_size_changed = 0;
-      intptr_t    msgbuf_size;
+      int64_t     msgbuf_size;
       //maybe a message?
       set_buf(&mpbuf, (u_char *)el->str, el->len);
       cmp_init(&cmp, &mpbuf, ngx_buf_reader, ngx_buf_writer);
@@ -1099,7 +1099,7 @@ static void redis_subscriber_callback(redisAsyncContext *c, void *r, void *privd
           }
           
           if(msgbuf_size_changed && (chanhead || ((chanhead = nchan_store_get_chanhead(chid, rdata)) != NULL))) {
-            chanhead->spooler.fn->broadcast_notice(&chanhead->spooler, NCHAN_NOTICE_REDIS_CHANNEL_MESSAGE_BUFFER_SIZE_CHANGE, (void *)msgbuf_size);
+            chanhead->spooler.fn->broadcast_notice(&chanhead->spooler, NCHAN_NOTICE_REDIS_CHANNEL_MESSAGE_BUFFER_SIZE_CHANGE, (void *)(intptr_t )msgbuf_size);
           }
           
           if(ngx_strmatch(&msg_type, "msg")) {
